[Automated tests] Remove the cleanup part from the extraction, it's generated based on the setup so we can do it in the runner
